Fixed discovery to skip NSE backed by invalid NSMgr
The Network Service Endpoint discovery function should not return
endpoints that were registered by those Network Service Managers that
are not part of the system anymore. If a stale endpoint is chosen, the
NS clients would not be able to connect to the NS even though other
valid endpoints exist in the system.